+++ packages/cmf/src/components/Saga/Saga.component.test.js
@@ -3,7 +3,6 @@ import { render } from '@testing-library/react';
 import { SagaComponent } from './Saga.component';
 
 const defaultMockUuid = '42';
-jest.mock('uuid', () => ({ v4: () => defaultMockUuid }));
 
 describe('Saga Component', () => {
 	it('should dispatch actions', () => {