+++ packages/components/src/TooltipTrigger/TooltipTrigger.component.js
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 import React, { cloneElement, useState, useRef } from 'react';
 import ReactDOM from 'react-dom';
-import uuid from 'uuid';
 import classNames from 'classnames';
+import { randomUUID } from '@talend/utils';
 import theme from './TooltipTrigger.module.scss';
 import useTooltipVisibility from './TooltipTrigger.hook';
 
@@ -112,7 +112,7 @@ function TooltipTrigger({
 
 	const [visible, show, hide] = useTooltipVisibility(tooltipDelay);
 
-	const [id] = useState(uuid.v4());
+	const [id] = useState(randomUUID());
 
 	const { props: childrenProps } = children;