+++ packages/datagrid/src/types/types.d.ts
@@ -4,3 +4,6 @@ declare module '*.scss' {
 }
 
 declare module '@talend/react-bootstrap';
+declare module '@talend/react-components' {
+	export * from '@talend/react-components/types';
+}